Delete unused build_project method and move build_library to test/runner.py. NFC.
build_library only had one non-test caller, and it turns out that
was itself dead code.
