+++ tools/shared.py
@@ -1699,111 +1699,6 @@ class Building(object):
       raise subprocess.CalledProcessError(cmd=args, returncode=res.returncode)
 
   @staticmethod
-  def build_library(name,
-                    build_dir,
-                    output_dir,
-                    generated_libs,
-                    configure=['sh', './configure'],
-                    configure_args=[],
-                    make=['make'],
-                    make_args='help',
-                    cache=None,
-                    cache_name=None,
-                    copy_project=False,
-                    env_init={},
-                    source_dir=None,
-                    native=False):
-    """Build a library into a .bc file. We build the .bc file once and cache it
-    for all our tests. (We cache in memory since the test directory is destroyed
-    and recreated for each test. Note that we cache separately for different
-    compilers).  This cache is just during the test runner. There is a different
-    concept of caching as well, see |Cache|.
-    """
-
-    if type(generated_libs) is not list:
-      generated_libs = [generated_libs]
-    if source_dir is None:
-      source_dir = path_from_root('tests', name.replace('_native', ''))
-    if make_args == 'help':
-      make_args = ['-j', str(multiprocessing.cpu_count())]
-
-    temp_dir = build_dir
-    if copy_project:
-      project_dir = os.path.join(temp_dir, name)
-      if os.path.exists(project_dir):
-        shutil.rmtree(project_dir)
-      shutil.copytree(source_dir, project_dir) # Useful in debugging sometimes to comment this out, and two lines above
-    else:
-      project_dir = build_dir
-    try:
-      old_dir = os.getcwd()
-    except:
-      old_dir = None
-    os.chdir(project_dir)
-    generated_libs = [os.path.join(project_dir, lib) for lib in generated_libs]
-    # for lib in generated_libs:
-    #   try:
-    #     os.unlink(lib) # make sure compilation completed successfully
-    #   except:
-    #     pass
-    env = Building.get_building_env(native, True)
-    for k, v in env_init.items():
-      env[k] = v
-    if configure:
-      # Useful in debugging sometimes to comment this out (and the lines below
-      # up to and including the |link| call)
-      if EM_BUILD_VERBOSE < 2:
-        stdout = open(os.path.join(project_dir, 'configure_out'), 'w')
-      else:
-        stdout = None
-      if EM_BUILD_VERBOSE < 1:
-        stderr = open(os.path.join(project_dir, 'configure_err'), 'w')
-      else:
-        stderr = None
-      try:
-        Building.configure(configure + configure_args, env=env, stdout=stdout, stderr=stderr)
-      except subprocess.CalledProcessError as e:
-        pass # Ignore exit code != 0
-
-    def open_make_out(i, mode='r'):
-      return open(os.path.join(project_dir, 'make_out' + str(i)), mode)
-
-    def open_make_err(i, mode='r'):
-      return open(os.path.join(project_dir, 'make_err' + str(i)), mode)
-
-    if EM_BUILD_VERBOSE >= 3:
-      make_args += ['VERBOSE=1']
-
-    # FIXME: Sad workaround for some build systems that need to be run twice to succeed (e.g. poppler)
-    for i in range(2):
-      with open_make_out(i, 'w') as make_out:
-        with open_make_err(i, 'w') as make_err:
-          stdout = make_out if EM_BUILD_VERBOSE < 2 else None
-          stderr = make_err if EM_BUILD_VERBOSE < 1 else None
-          try:
-            Building.make(make + make_args, stdout=stdout, stderr=stderr, env=env)
-          except subprocess.CalledProcessError as e:
-            pass # Ignore exit code != 0
-      try:
-        if cache is not None:
-          cache[cache_name] = []
-          for f in generated_libs:
-            basename = os.path.basename(f)
-            cache[cache_name].append((basename, open(f, 'rb').read()))
-        break
-      except Exception as e:
-        if i > 0:
-          if EM_BUILD_VERBOSE == 0:
-            # Due to the ugly hack above our best guess is to output the first run
-            with open_make_err(0) as ferr:
-              for line in ferr:
-                sys.stderr.write(line)
-          raise Exception('could not build library ' + name + ' due to exception ' + str(e))
-    if old_dir:
-      os.chdir(old_dir)
-    return generated_libs
-
-  @staticmethod
   def make_paths_absolute(f):
     if f.startswith('-'):  # skip flags
       return f
